+++ zhglxt-quartz/src/main/java/com/zhglxt/quartz/controller/SysJobLogController.java
@@ -1,6 +1,7 @@
 package com.zhglxt.quartz.controller;
 
 import com.zhglxt.common.annotation.Log;
+import com.zhglxt.common.config.GlobalConfig;
 import com.zhglxt.common.core.controller.BaseController;
 import com.zhglxt.common.core.entity.AjaxResult;
 import com.zhglxt.common.core.page.TableDataInfo;
@@ -69,6 +70,9 @@ public class SysJobLogController extends BaseController {
     @PostMapping("/remove")
     @ResponseBody
     public AjaxResult remove(String ids) {
+        if (GlobalConfig.isDemoEnabled()) {
+            return error("演示模式不允许本操作");
+        }
         return toAjax(jobLogService.deleteJobLogByIds(ids));
     }
 
@@ -85,6 +89,9 @@ public class SysJobLogController extends BaseController {
     @PostMapping("/clean")
     @ResponseBody
     public AjaxResult clean() {
+        if (GlobalConfig.isDemoEnabled()) {
+            return error("演示模式不允许本操作");
+        }
         jobLogService.cleanJobLog();
         return success();
     }
